独立debug和config代码，保持主文件清洁
