+++ .gitignore
@@ -2,3 +2,4 @@ test/
 pic/
 screenshot_backups/
 autojump.png
+*.pyc
