+++ wechat_jump_auto.py
@@ -2,14 +2,11 @@
 import os
 import sys
 import subprocess
-import shutil
 import time
 import math
-from PIL import Image, ImageDraw
+from PIL import Image
 import random
-import json
-import re
-
+from WechatJump import Debug, Config
 
 # === 思路 ===
 # 核心：每次落稳之后截图，根据截图算出棋子的坐标和下一个块顶面的中点坐标，
@@ -22,39 +19,13 @@ import re
 #      这时候得到了块中点的 X 轴坐标，这时候假设现在棋子在当前块的中心，
 #      根据一个通过截图获取的固定的角度来推出中点的 Y 坐标
 # 最后：根据两点的坐标算距离乘以系数来获取长按时间（似乎可以直接用 X 轴距离）
-
-
 # TODO: 解决定位偏移的问题
 # TODO: 看看两个块中心到中轴距离是否相同，如果是的话靠这个来判断一下当前超前还是落后，便于矫正
 # TODO: 一些固定值根据截图的具体大小计算
 # TODO: 直接用 X 轴距离简化逻辑
 
-def open_accordant_config():
-    screen_size = _get_screen_size()
-    config_file = "{path}/config/{screen_size}/config.json".format(
-        path=sys.path[0],
-        screen_size=screen_size
-    )
-    if os.path.exists(config_file):
-        with open(config_file, 'r') as f:
-            print("Load config file from {}".format(config_file))
-            return json.load(f)
-    else:
-        with open('{}/config/default.json'.format(sys.path[0]), 'r') as f:
-            print("Load default config")
-            return json.load(f)
-
-
-def _get_screen_size():
-    size_str = os.popen('adb shell wm size').read()
-    m = re.search('(\d+)x(\d+)', size_str)
-    if m:
-        width = m.group(1)
-        height = m.group(2)
-        return "{height}x{width}".format(height=height, width=width)
-
-
-config = open_accordant_config()
+debug_switch = False # debug开关，需要调试的时候请改为：True
+config = Config.open_accordant_config()
 
 # Magic Number，不设置可能无法正常执行，请根据具体截图从上到下按需设置
 under_game_score_y = config['under_game_score_y']
@@ -69,12 +40,6 @@ else:
     swipe = {}
     swipe['x1'], swipe['y1'], swipe['x2'], swipe['y2'] = 320, 410, 320, 410
 
-
-screenshot_backup_dir = 'screenshot_backups/'
-if not os.path.isdir(screenshot_backup_dir):
-        os.mkdir(screenshot_backup_dir)
-
-
 def pull_screenshot():
     process = subprocess.Popen('adb shell screencap -p', shell=True, stdout=subprocess.PIPE)
     screenshot = process.stdout.read()
@@ -83,26 +48,6 @@ def pull_screenshot():
     f = open('autojump.png', 'wb')
     f.write(screenshot)
     f.close()
-
-def backup_screenshot(ts):
-    # 为了方便失败的时候 debug
-    if not os.path.isdir(screenshot_backup_dir):
-        os.mkdir(screenshot_backup_dir)
-    shutil.copy('autojump.png', '{}{}.png'.format(screenshot_backup_dir, ts))
-
-
-def save_debug_creenshot(ts, im, piece_x, piece_y, board_x, board_y):
-    draw = ImageDraw.Draw(im)
-    # 对debug图片加上详细的注释
-    draw.line((piece_x, piece_y) + (board_x, board_y), fill=2, width=3)
-    draw.line((piece_x, 0, piece_x, im.size[1]), fill=(255, 0, 0))
-    draw.line((0, piece_y, im.size[0], piece_y), fill=(255, 0, 0))
-    draw.line((board_x, 0, board_x, im.size[1]), fill=(0, 0, 255))
-    draw.line((0, board_y, im.size[0], board_y), fill=(0, 0, 255))
-    draw.ellipse((piece_x - 10, piece_y - 10, piece_x + 10, piece_y + 10), fill=(255, 0, 0))
-    draw.ellipse((board_x - 10, board_y - 10, board_x + 10, board_y + 10), fill=(0, 0, 255))
-    del draw
-    im.save('{}{}_d.png'.format(screenshot_backup_dir, ts))
 
 
 def set_button_position(im):
@@ -223,12 +168,13 @@ def main():
         im = Image.open('./autojump.png')
         # 获取棋子和 board 的位置
         piece_x, piece_y, board_x, board_y = find_piece_and_board(im)
-        ts = int(time.time())
-        print(ts, piece_x, piece_y, board_x, board_y)
         set_button_position(im)
         jump(math.sqrt((board_x - piece_x) ** 2 + (board_y - piece_y) ** 2))
-        save_debug_creenshot(ts, im, piece_x, piece_y, board_x, board_y)
-        backup_screenshot(ts)
+        if debug_switch:
+            ts = int(time.time())
+            print(ts, piece_x, piece_y, board_x, board_y)
+            Debug.backup_screenshot(ts)
+            Debug.save_debug_screenshot(ts, im, piece_x, piece_y, board_x, board_y)
         time.sleep(random.uniform(1, 1.1))   # 为了保证截图的时候应落稳了，多延迟一会儿
